pyproject.toml
[tool.poetry]
name = "morpheus"
version = "0.1.0"
description = "ZEUS OS Test Framework"
authors = ["Pedro Thaines <pedro@citeb.com.br>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
robotframework = "^4.1.3"
robotframework-seleniumlibrary = "^5.1.3"
robotframework-robocop = "^1.12.0"
robotframework-requests = "^0.9.2"
robotframework-tidy = "^1.6.2"
robotframework-sshlibrary = "^3.8.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
